app/src/main/java/id/ariefpurnamamuharram/myfootballapplication/teamdetails/TeamDetailsPresenter.kt
package id.ariefpurnamamuharram.myfootballapplication.teamdetails

import com.google.gson.Gson
import id.ariefpurnamamuharram.myfootballapplication.api.ApiRepository
import id.ariefpurnamamuharram.myfootballapplication.api.TheSportDBApi
import id.ariefpurnamamuharram.myfootballapplication.model.TeamDetailsResponse
import id.ariefpurnamamuharram.myfootballapplication.util.CoroutineContextProvider
import id.ariefpurnamamuharram.myfootballapplication.view.LoadingView
import kotlinx.coroutines.experimental.async
import org.jetbrains.anko.coroutines.experimental.bg

class TeamDetailsPresenter(private val view: LoadingView,
                           private val content: TeamDetailsView,
                           private val apiRepository: ApiRepository,
                           private val gson: Gson,
                           private val context: CoroutineContextProvider = CoroutineContextProvider()) {

    fun getTeamDetails(name: String?) {
        view.showLoading()

        async(context.main) {
            val data = bg {
                gson.fromJson(apiRepository
                        .doRequest(TheSportDBApi.getTeamDetails(name)),
                        TeamDetailsResponse::class.java)
            }

            content.showTeamDetails(data.await().teamDetails)
            view.hideLoading()
        }
    }
}